Implemented TryFrom and Drop for XRTDevice, renaming, added tests
