+++ README.md
@@ -1,0 +1,7 @@
+# xrt-rs
+In case linking fails, add the XRT libs to your library path:
+```
+export LD_LIBRARY_PATH="$LD_LIBRARY_PATH:/opt/xilinx/xrt/lib"
+or
+export LD_LIBRARY_PATH="$LD_LIBRARY_PATH:$XILINX_XRT"
+```
